Rename `util_iter` module to `read_iters`.
